+++ src/parsley/validator.js
@@ -136,17 +136,18 @@ define('parsley/validator', [
         return $.extend(assert, { priority: 256 });
       },
       pattern: function (regexp) {
-        var flags = null;
-        var hasFlags = regexp.match(/^\/(.*)\/([gimy]+)$/);
-        if (hasFlags) {
-          regexp = hasFlags[1];
-          flags = hasFlags[2];
-        } else {
-          var enclosed = regexp.match(/^\/(.*)\/$/);
-          if (enclosed) {
-            regexp = enclosed[1];
-          }
+        var flags = '';
+
+        // Test if RegExp is literal, if not, nothing to be done, otherwise, we need to isolate flags and pattern
+        if (!!(/^\/.*\/(?:[gimy]*)$/.test(regexp))) {
+          // Replace the regexp literal string with the first match group: ([gimy]*)
+          // If no flag is present, this will be a blank string
+          flags = regexp.replace(/.*\/([gimy]*)$/, '$1');
+          // Again, replace the regexp literal string with the first match group:
+          // everything excluding the opening and closing slashes and the flags
+          regexp = regexp.replace(new RegExp('^/(.*?)/' + flags + '$'), '$1');
         }
+
         return $.extend(new Validator.Assert().Regexp(regexp, flags), { priority: 64 });
       },
       minlength: function (value) {